backend/flask_project/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pymysql==1.1.1
sqlalchemy==2.0.39
python-dotenv==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.9.2
email-validator==2.2.0
fastapi-mail==1.4.1
python-multipart==0.0.9